pyproject.toml
[project]
name = "journeyiq"
version = "0.1.0"
description = "Itilite JourneyIQ — AI-powered corporate travel chat backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.26.0",
    "fastapi>=0.115.0",
    "python-socketio>=5.11.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.13.3",
]
